+++ skrypty/ui/gags/color/color_zaslony_nieudane.lua
@@ -1,4 +1,8 @@
 function trigger_func_skrypty_ui_gags_color_color_zaslony_nieudane_nie_zaslaniasz_ty()
+    if scripts.gags:delete_line("zaslony_nieudane") then
+        return
+    end
+
     selectCurrentLine()
     deleteLine()
     cecho("\n<" .. scripts.gag_colors["zaslony_nieudane"] .. ">[N ZASLANIASZ] " .. matches[2] .. "\n")
@@ -6,6 +10,10 @@ function trigger_func_skrypty_ui_gags_color_color_zaslony_nieudane_nie_zaslanias
 end
 
 function trigger_func_skrypty_ui_gags_color_color_zaslony_nieudane_nie_zaslania_kogos()
+    if scripts.gags:delete_line("zaslony_nieudane") then
+        return
+    end
+
     selectCurrentLine()
     deleteLine()
     cecho("\n<" .. scripts.gag_colors["zaslony_nieudane"] .. ">[ N ZASLANIA ] " .. matches[2] .. "\n")
@@ -13,6 +21,10 @@ function trigger_func_skrypty_ui_gags_color_color_zaslony_nieudane_nie_zaslania_
 end
 
 function trigger_func_skrypty_ui_gags_color_color_zaslony_nieudane_nie_wycofanie_ty()
+    if scripts.gags:delete_line("zaslony_nieudane") then
+        return
+    end
+    
     selectCurrentLine()
     local str_replace = "[ NIE WYCOFU ] "
     prefix(str_replace)
